requirements.txt
agno
pytest
pytest-asyncio
pytest-cov
python-dotenv
requests
pydantic
sqlalchemy>=2.0
psycopg2-binary
alembic
bcrypt
fastapi
uvicorn[standard]
pyjwt
email-validator
python-multipart
pytest-mock
